feat:增加门禁控制接口

+++ app/src/main/java/com/meeting/accesscontrol/bean/JudgeRequest.kt
@@ -1,6 +1,6 @@
 package com.meeting.accesscontrol.bean
 
 data class JudgeRequest(
-    val userId: Int,    //用户ID 编号
-    val meetingId: Int  //会议Id
+    val userId: Long,    //用户ID 编号
+    val meetingId: Long  //会议Id
 )
